coninue ts conversion and fix google font
