+++ package.json
@@ -41,6 +41,7 @@
     "@types/node": "^24.5.1",
     "@types/react": "^19.1.13",
     "@types/react-dom": "^19.1.9",
+    "typed-css-modules": "^0.9.1",
     "typescript": "^4.9.5"
   },
   "type": "module"
