+++ public/index.html
@@ -24,10 +24,21 @@
       work correctly both with client-side routing and a non-root public URL.
       Learn how to configure a non-root public URL by running `npm run build`.
     -->
-    <title>React App</title>
+    <!-- Google fonts -->
+    <link rel="preconnect" href="https://fonts.googleapis.com"></link>
+    <link
+      rel="preconnect"
+      href="https://fonts.gstatic.com"
+      crossOrigin=""
+    ></link>
+    <link
+      href="https://fonts.googleapis.com/css2?family=New+Rocker&display=swap"
+      rel="stylesheet"
+    ></link>
+    <title>YarnCalc</title>
   </head>
   <body>
-    <noscript>You need to enable JavaScript to run this app.</noscript>
+    <noscript>Helpful crafting calculators</noscript>
     <div id="root"></div>
     <!--
       This HTML file is a template.
